+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.23.3](https://github.com/valentine195/obsidian-5e-statblocks/compare/2.23.2...2.23.3) (2022-12-07)
+
+
+### Bug Fixes
+
+* fixes dropzone inheritance issue ([8e10905](https://github.com/valentine195/obsidian-5e-statblocks/commit/8e1090511e02cb21067cee26a321430d2ed2f5db))
+
+### [2.23.2](https://github.com/valentine195/obsidian-5e-statblocks/compare/2.23.1...2.23.2) (2022-12-06)
+
+
+### Bug Fixes
+
+* improves subheading editing experience ([7764e51](https://github.com/valentine195/obsidian-5e-statblocks/commit/7764e5107c71edf1217b0eb814c1494594917dda))
+* removes logs ([6b0d7ea](https://github.com/valentine195/obsidian-5e-statblocks/commit/6b0d7ea14aa3a540346d034500219b2cd7a9a063))
+
+### [2.23.1](https://github.com/valentine195/obsidian-5e-statblocks/compare/2.23.0...2.23.1) (2022-12-06)
+
+
+### Bug Fixes
+
+* empty inline groups easier to drop ([20744a0](https://github.com/valentine195/obsidian-5e-statblocks/commit/20744a036864117dad50bc4a3242137ee5ee0bdf))
+* fixes nested conditioned property behavior (closes [#116](https://github.com/valentine195/obsidian-5e-statblocks/issues/116)) ([6fc2aee](https://github.com/valentine195/obsidian-5e-statblocks/commit/6fc2aee9c8ec605600bc2c7b43120860e03da7de))
+* fixes nested inline content not displayed (close [#105](https://github.com/valentine195/obsidian-5e-statblocks/issues/105)) ([4c50bf0](https://github.com/valentine195/obsidian-5e-statblocks/commit/4c50bf040058a76c003429a46bfbb3034ac04a40))
+
 ## [2.23.0](https://github.com/valentine195/obsidian-5e-statblocks/compare/2.22.6...2.23.0) (2022-11-23)